+++ src/tanok.js
@@ -1,29 +1,4 @@
-import {tanok, effectWrapper, TanokDispatcher} from './core';
-import {tanokComponent} from './component';
-import {on} from './decorators';
-import {
-  actionIs,
-  parentIs,
-  filter,
-  debounce,
-  throttle,
-  rethrowFx,
-  subcomponentFx,
-} from './helpers';
-
-export {
-  tanok,
-  effectWrapper,
-  TanokDispatcher,
-  on,
-
-  actionIs,
-  parentIs,
-  filter,
-  debounce,
-  throttle,
-  rethrowFx,
-  subcomponentFx,
-
-  tanokComponent,
-};
+export * from './core';
+export * from './component';
+export * from './decorators';
+export * from './helpers';