+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## 1.0
 
+* **BREAKING** Field `StreamWrapper.parent` was renamed to `StreamWrapper.streamName` in  and in messages that `StreamWrapper.send` pushes to stream
 * Updated rollup version